Add files via upload
SLIDING LED

Hello everyone,

We are together with a new project, this time it will turn on and off 5 LEDs in sequence without a microcontroller or a password-integrating IC. You can use it for a signal indicator or a bicycle. With this timing, we will charge a capacitor in front of us, and with the effect of this charge, our transfers will trigger in turn, and this place connected to them will start to burn, again in the same way, when the signal is interrupted, our capacitor will be Discharged, this time our flavors will continue in order In short, our revolution simply works like this.